project1/python_ver/requirements.txt
cryptography==38.0.3
pycryptodome==3.15.0
